+++ include/net/http_parser.h
@@ -46,6 +46,8 @@ typedef unsigned __int64 u64_t;
 #include <zephyr/types.h>
 #include <stddef.h>
 #endif
+#include <net/http_parser_state.h>
+#include <net/http_parser_url.h>
 
 /* Maximium header size allowed. If the macro is not defined
  * before including this header then the default is used. To
@@ -229,38 +231,6 @@ struct http_parser_settings {
 };
 
 
-enum http_parser_url_fields {
-	  UF_SCHEMA           = 0
-	, UF_HOST             = 1
-	, UF_PORT             = 2
-	, UF_PATH             = 3
-	, UF_QUERY            = 4
-	, UF_FRAGMENT         = 5
-	, UF_USERINFO         = 6
-	, UF_MAX              = 7
-};
-
-
-/* Result structure for http_parser_parse_url().
- *
- * Callers should index into field_data[] with UF_* values iff field_set
- * has the relevant (1 << UF_*) bit set. As a courtesy to clients (and
- * because we probably have padding left over), we convert any port to
- * a u16_t.
- */
-struct http_parser_url {
-	u16_t field_set;           /* Bitmask of (1 << UF_*) values */
-	u16_t port;                /* Converted UF_PORT string */
-
-	struct {
-		u16_t off;               /* Offset into buffer in which field
-					     * starts
-					     */
-		u16_t len;               /* Length of run in buffer */
-	} field_data[UF_MAX];
-};
-
-
 /* Returns the library version. Bits 16-23 contain the major version number,
  * bits 8-15 the minor version number and bits 0-7 the patch level.
  * Usage example:
@@ -305,13 +275,6 @@ const char *http_errno_name(enum http_errno err);
 
 /* Return a string description of the given error */
 const char *http_errno_description(enum http_errno err);
-
-/* Initialize all http_parser_url members to 0 */
-void http_parser_url_init(struct http_parser_url *u);
-
-/* Parse a URL; return nonzero on failure */
-int http_parser_parse_url(const char *buf, size_t buflen,
-			  int is_connect, struct http_parser_url *u);
 
 /* Pause or un-pause the parser; a nonzero value pauses */
 void http_parser_pause(struct http_parser *parser, int paused);
