+++ subsys/net/lib/http/Makefile
@@ -1,5 +1,6 @@
 ccflags-$(CONFIG_HTTP_PARSER_STRICT) += -DHTTP_PARSER_STRICT
 
 obj-$(CONFIG_HTTP_PARSER) := http_parser.o
+obj-$(CONFIG_HTTP_PARSER_URL) += http_parser_url.o
 obj-$(CONFIG_HTTP_CLIENT) += http_client.o
 obj-$(CONFIG_HTTP_SERVER) += http_server.o
