+++ subsys/net/lib/http/README_http_parser
@@ -8,4 +8,7 @@ can be found at:
 https://github.com/nodejs/http-parser/releases/tag/v2.7.1
 https://github.com/nodejs/http-parser/archive/v2.7.1.tar.gz
 
+NOTE: The portions which relate to URL parsing have been split out into
+http_parser_url.c (orignially located in http_parser.c).
+
 * "http-parser" is the project's name, "http_parser" is used in filenames.
